requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
Markdown==3.4.3
psycopg2==2.9.5
python-decouple==3.8
pytz==2023.2
redis==4.5.3
sqlparse==0.4.3
